Fix slider position filtering
The onInputReport is firing twice for single button presses, which
messed up the slider position filtering logic. Previously, we would
filter out these duplicates, but now that we filter the slider
positions, the second event would actually show up as a changed value
and therefore also trigger two events with different button events.

Easy solution is to apply the duplicate checking before doing the
filtering, which is more in line with the previous implementation.